backend/requirements.txt
# Web Framework
flask>=3.0.0
flask-sock>=0.7.0
flask-socketio>=5.3.0
flask-cors>=4.0.0
simple-websocket>=1.0.0

# Computer Vision
opencv-python>=4.10.0
numpy>=1.26.0
pillow>=10.1.0

# Hardware Communication
pyserial>=3.5

# Audio
pygame>=2.6.0

# ML
requests>=2.31.0
python-dotenv>=1.0.0

# YOLO for local test
ultralytics>=8.3.0  # YOLOv10 support

# Material Classification (M1-optimized)
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0

# Utilities
scipy==1.11.4
